third_party/WebKit/LayoutTests/http/tests/devtools/console/console-links-in-errors-with-trace-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test that relative links and links with hash open in the sources panel.


Running: testClickRelativeLink
Clicking link source3.js:3
Panel sources was opened.

Running: testClickURLWithHash
Panel console was opened.
Clicking link http://127.0.0.1:8000/devtools/resources/source3.js#foo
Panel sources was opened.
